Panier deconnecte

+++ project/views/commander.py
@@ -1,8 +1,9 @@
 from datetime import datetime
 
 from flask_sqlalchemy import SQLAlchemy
+import sqlalchemy
 from project import app, db
-from flask import flash, render_template, url_for, redirect, request, make_response
+from flask import flash, render_template, url_for, redirect, request, make_response, session
 from flask_wtf import FlaskForm
 from wtforms.validators import DataRequired
 from wtforms import BooleanField, SubmitField, StringField
@@ -10,7 +11,25 @@ from wtforms import HiddenField, IntegerField
 from wtforms.validators import DataRequired
 from flask_login import login_user , current_user, logout_user, login_required
 from hashlib import sha256
-from project.models import Plats, Allergenes, Constituer, Commandes, Formule
+from project.models import Plats, Allergenes, Constituer, Commandes, Formule, User, UserType
+
+def get_current_user():
+    if current_user.is_authenticated:
+        return current_user
+    elif session.get('user') is not None:
+        user_id = session.get('user')
+        return User.get_user(user_id)
+    return None
+
+def callback_type_user(callback_user, callback_guest, callback_unknown):
+    if current_user.is_authenticated:
+        return callback_user()
+    elif session.get('user') is not None:
+        return callback_guest()
+    return callback_unknown()
+
+def get_type_user():
+    return callback_type_user(lambda: UserType.USER, lambda: UserType.GUEST, lambda: UserType.UNKNOW)
 
 class CommanderForm(FlaskForm):
     id_plat = HiddenField()
@@ -19,42 +38,46 @@ class CommanderForm(FlaskForm):
 
 @app.route("/commander", methods=["GET", "POST"])
 def commander():
-    if current_user:
-        commande = current_user.get_or_create_panier()
-        num_commande = commande.num_commande
-        form = CommanderForm()
-        selected_allergenes = request.form.getlist('allergenes')  # Liste des allergènes cochés
-        type = request.args.get('type', 'p')
-        allergenes = Allergenes.get_allergenes()
-        plats = Plats.get_plats()
-        plats_chauds = Plats.get_plats_filtered_by_type_and_allergenes("Plat chaud", selected_allergenes)
-        plats_froids = Plats.get_plats_filtered_by_type_and_allergenes("Plat froid", selected_allergenes)
-        sushis = Plats.get_plats_filtered_by_type_and_allergenes("Sushi", selected_allergenes)
-        formules = Formule.get_formules_filtered_by_allergenes(selected_allergenes)
-        desserts = Plats.get_plats_filtered_by_type_and_allergenes("Dessert", selected_allergenes)
-
-        return render_template("commander.html", 
-                            plats=plats, 
-                            plats_chauds=plats_chauds,
-                            plats_froids=plats_froids,
-                            sushis=sushis,
-                            formules=formules, 
-                            desserts=desserts, 
-                            type=type, 
-                            nb_plats=len(plats), 
-                            nb_formules=len(formules), 
-                            nb_desserts=len(desserts), 
-                            allergenes=allergenes, 
-                            selected_allergenes=selected_allergenes,
-                            form=form,
-                            num_com = num_commande)
+    user = get_current_user()
+    if user is None:
+        return redirect(url_for('login_unsafe'))
     
+    commande = user.get_or_create_panier()
+    num_commande = commande.num_commande
+    form = CommanderForm()
+    selected_allergenes = request.form.getlist('allergenes')  # Liste des allergènes cochés
+    type = request.args.get('type', 'p')
+    allergenes = Allergenes.get_allergenes()
+    plats = Plats.get_plats()
+    plats_chauds = Plats.get_plats_filtered_by_type_and_allergenes("Plat chaud", selected_allergenes)
+    plats_froids = Plats.get_plats_filtered_by_type_and_allergenes("Plat froid", selected_allergenes)
+    sushis = Plats.get_plats_filtered_by_type_and_allergenes("Sushi", selected_allergenes)
+    formules = Formule.get_formules_filtered_by_allergenes(selected_allergenes)
+    desserts = Plats.get_plats_filtered_by_type_and_allergenes("Dessert", selected_allergenes)
+
+    return render_template("commander.html", 
+                        plats=plats, 
+                        plats_chauds=plats_chauds,
+                        plats_froids=plats_froids,
+                        sushis=sushis,
+                        formules=formules, 
+                        desserts=desserts, 
+                        type=type, 
+                        nb_plats=len(plats), 
+                        nb_formules=len(formules), 
+                        nb_desserts=len(desserts), 
+                        allergenes=allergenes, 
+                        selected_allergenes=selected_allergenes,
+                        form=form,
+                        num_com = num_commande)
+
 @app.route("/filter_allergenes", methods=["GET", "POST"])
 def filter_allergenes():
 
-    if not current_user:
-        return redirect(url_for('commander'))
-    
+    user = get_current_user()
+    if user is None:
+        return redirect(url_for('login_unsafe'))
+        
     # Récupérer la liste des allergènes sélectionnés
     if request.method == "POST":
         selected_allergenes = request.form.getlist('allergenes') # Liste des allergènes cochés
@@ -83,7 +106,7 @@ def filter_allergenes():
     desserts = Plats.get_plats_filtered_by_type_and_allergenes("Dessert", selected_allergenes)
     
 
-    commande = current_user.get_or_create_panier()
+    commande = user.get_or_create_panier()
     num_commande = commande.num_commande
     form = CommanderForm()
 
@@ -112,31 +135,31 @@ def filter_allergenes():
 
 @app.route("/commander_plat", methods = ("POST",))
 def ajout_plat() :
-    if current_user :
-        f = CommanderForm()
-        if f.num_com.data :
-            try:
-                commande = Commandes.get_commande(f.num_com.data)
-                constituer = Constituer.get_constituer(f.id_plat.data, f.num_com.data)
-                if constituer:
-                    constituer.quantite_plat += f.quantite.data
-                else:
-                    constituer = Constituer(id_plat = f.id_plat.data, num_commande = f.num_com.data, quantite_plat = f.quantite.data)
-                    commande.constituer_assoc.append(constituer)
-                db.session.add(constituer)
-                db.session.commit()
-            except Exception as e:
-                flash("Erreur : " + str(e._message), "danger")
-                return redirect(url_for('commander'))
+    f = CommanderForm()
+    if f.num_com.data :
+        try:
+            commande = Commandes.get_commande(f.num_com.data)
+            constituer = Constituer.get_constituer(f.id_plat.data, f.num_com.data)
+            if constituer:
+                constituer.quantite_plat += f.quantite.data
+            else:
+                constituer = Constituer(id_plat = f.id_plat.data, num_commande = f.num_com.data, quantite_plat = f.quantite.data)
+                commande.constituer_assoc.append(constituer)
+            db.session.add(constituer)
+            db.session.commit()
+        except Exception as e:
+            flash("Erreur : " + str(e.orig.args[1]), "danger")
+            return redirect(url_for('commander'))
 
     return redirect(url_for('commander'))
 
 @app.route("/panier")
 def panier():
-    if not current_user.is_authenticated:
-        return redirect(url_for('login'))
+    user = get_current_user()
+    if user is None:
+        return redirect(url_for('login_unsafe'))
     
-    panier = current_user.get_or_create_panier()
+    panier = user.get_or_create_panier()
     if panier is not None:
         panier.calculer_prix()
         panier.compute_reduction()
@@ -151,20 +174,18 @@ def panier():
 def modifier_quantite():
     action = request.args.get('action')
     nom_plat = request.args.get('nom_plat')
-    if not current_user.is_authenticated:
-        panier = None
-    else:
-        panier = current_user.get_panier()
-
-    if panier is not None:
-        for constituer in panier.constituer_assoc:
-            if constituer.plat.nom_plat == nom_plat:
-                if action == 'increment':
-                    if constituer.quantite_plat +1 <= int(constituer.plat.stock_utilisable * 0.8):
-                        constituer.quantite_plat += 1
-                elif action == 'decrement' and constituer.quantite_plat > 1:
-                        constituer.quantite_plat -= 1
-                break
+    user = get_current_user()
+    if user is not None:
+        panier = user.get_panier()
+        if panier is not None:
+            for constituer in panier.constituer_assoc:
+                if constituer.plat.nom_plat == nom_plat:
+                    if action == 'increment':
+                        if constituer.quantite_plat +1 <= int(constituer.plat.stock_utilisable * 0.8):
+                            constituer.quantite_plat += 1
+                    elif action == 'decrement' and constituer.quantite_plat > 1:
+                            constituer.quantite_plat -= 1
+                    break
 
         db.session.commit()
 
@@ -173,51 +194,49 @@ def modifier_quantite():
 @app.route('/modifier_date_heure')
 def modifier_date_heure():
     hours = request.args.get('datetime')
-    if not current_user.is_authenticated:
-        panier = None
-    else:
-        panier = current_user.get_panier()
 
-    if panier is not None:
-        today = datetime.today()
-        time = datetime.strptime(hours, "%H:%M").time()
-        panier.date = datetime.combine(today, time)
-        try:
-            db.session.commit()
-        except Exception as e:
-            flash("Erreur : " + str(e._message), "danger")
-            return redirect(url_for('panier'))
+    user = get_current_user()
+    if user is not None:
+        panier = user.get_panier()
+        if panier is not None:
+            today = datetime.today()
+            time = datetime.strptime(hours, "%H:%M").time()
+            panier.date = datetime.combine(today, time)
+            try:
+                db.session.commit()
+            except sqlalchemy.exc.OperationalError as e:
+                flash("Erreur : " + str(e.orig.args[1]), "danger")
+                return redirect(url_for('panier'))
 
     return redirect(url_for('panier'))
 
 @app.route('/modifier_type')
 def modifier_type():
     sur_place = request.args.get('delivery')
-    if not current_user.is_authenticated:
-        panier = None
-    else:
-        panier = current_user.get_panier()
-
-    if panier is not None:
-        if sur_place == "1":
-            numero_table = Commandes.get_num_table_dispo(panier.date)
-            if numero_table != -1:
-                if panier.date is not None and panier.date.time() > datetime.strptime("14:00", "%H:%M").time():
-                    panier.date = datetime.combine(panier.date, datetime.strptime("13:50", "%H:%M").time())
-                panier.sur_place = True
-                panier.num_table = numero_table
+    
+    user = get_current_user()
+    if user is not None:
+        panier = user.get_panier()
+        if panier is not None:
+            if sur_place == "1":
+                numero_table = Commandes.get_num_table_dispo(panier.date)
+                if numero_table != -1:
+                    if panier.date is not None and panier.date.time() > datetime.strptime("14:00", "%H:%M").time():
+                        panier.date = datetime.combine(panier.date, datetime.strptime("13:50", "%H:%M").time())
+                    panier.sur_place = True
+                    panier.num_table = numero_table
+                else:
+                    panier.sur_place = False 
+                    panier.num_table = None
             else:
-                panier.sur_place = False 
+                panier.sur_place = False
                 panier.num_table = None
-        else:
-            panier.sur_place = False
-            panier.num_table = None
-            
-        try:
-            db.session.commit()
-        except Exception as e:
-            flash("Erreur : " + str(e._message), "danger")
-            return redirect(url_for('panier'))
+                
+            try:
+                db.session.commit()
+            except Exception as e:
+                flash("Erreur : " + str(e.orig.args[1]), "danger")
+                return redirect(url_for('panier'))
 
 
     return redirect(url_for('panier'))
@@ -226,17 +245,13 @@ def modifier_type():
 @app.route('/supprimer_plat')
 def supprimer_plat():
     nom_plat = request.args.get('nom_plat')
-    if not current_user.is_authenticated:
-        panier = None
-    else:
-        panier = current_user.get_panier()
+    user = get_current_user()
+    if user is not None:
+        for constituer in user.get_panier().constituer_assoc:
+            if constituer.plat.nom_plat == nom_plat:
+                db.session.delete(constituer)
 
-    for constituer in panier.constituer_assoc:
-        if constituer.plat.nom_plat == nom_plat:
-            db.session.delete(constituer)
-
-    db.session.commit()
-
+        db.session.commit()
     return redirect(url_for('panier'))
 
 @app.route("/choix_paiement")
@@ -250,11 +265,11 @@ def paiement_cb():
 
 @app.route("/paiement/validation", methods = ["POST"])
 def validation_paiement():
-    if not current_user.is_authenticated:
-        panier = None
-    else:
-        panier = current_user.get_panier()
-
+    user = get_current_user()
+    if user is None:
+        return redirect(url_for('login'))
+                                
+    panier = user.get_panier()
     if panier is None:
         return redirect(url_for('panier'))
 
@@ -268,10 +283,7 @@ def validation_paiement():
         db.session.commit()
 
     except Exception as e:
-        flash("Erreur : " + str(e._message), "danger")
+        flash("Erreur : " + str(e.orig.args[1]), "danger")
         return redirect(url_for('panier'))
-
-
-
 
     return render_template("validation_commande.html", panier=panier)
